fix: Generate d.ts files
