+++ src/types.ts
@@ -1,4 +1,4 @@
-import { Result } from 'axe-core'
+import { Result, Check, ImpactValue, Locale, Rule, RunOptions } from 'axe-core'
 
 export interface NodeViolation {
   target: string
@@ -17,3 +17,34 @@ export interface Aggregate {
 export default interface Reporter {
   report(violations: Result[]): Promise<void>
 }
+
+export interface axeOptionsConfig {
+  axeOptions?: RunOptions
+}
+
+export interface ConfigOptions {
+  branding?: {
+    brand?: string
+    application?: string
+  }
+  reporter?: 'v1' | 'v2' | 'no-passes'
+  checks?: Check[]
+  rules?: Rule[]
+  locale?: Locale
+  axeVersion?: string
+}
+
+/**
+ * Implement this interface to be able to specific custom reporting behaviour for checkA11y method.
+ * @see checkA11y
+ */
+export default interface Reporter {
+  report(violations: Result[]): Promise<void>
+}
+
+export type Options = {
+  includedImpacts?: ImpactValue[]
+  detailedReport?: boolean
+  detailedReportOptions?: { html?: boolean }
+  verbose?: boolean
+} & axeOptionsConfig
