+++ tsconfig.json
@@ -3,7 +3,7 @@
     "target": "es6",
     "module": "commonjs",
     "strict": true,
-    "declaration": false,
+    "declaration": true,
     "rootDir": "src",
     "moduleResolution": "node",
     "typeRoots": ["node_modules/@types"],
